+++ src/moduls/ui/VCAEngine/project.cpp
@@ -911,7 +911,7 @@ string Page::calcProg( )
     if(!mProc.size() && !parent().freeStat()) return parent().at().calcProg();
 
     string iprg = mProc;
-    unsigned lng_end = iprg.find("\n");
+    size_t lng_end = iprg.find("\n");
     if(lng_end == string::npos) lng_end = 0;
     else lng_end++;
     return iprg.substr(lng_end);
