+++ src/moduls/ui/WebCfgD/web_cfg.cpp
@@ -493,7 +493,7 @@ string TWEB::getCookie( string name, vector<string> &vars )
     for( unsigned i_var = 0; i_var < vars.size(); i_var++)
 	if( vars[i_var].substr(0, vars[i_var].find(":",0)) == "Cookie" )
 	{
-	    unsigned i_beg = vars[i_var].find(name+"=",0);
+	    size_t i_beg = vars[i_var].find(name+"=",0);
 	    if( i_beg == string::npos ) return "";
 	    i_beg += name.size()+1;
 	    return vars[i_var].substr(i_beg,vars[i_var].find(";",i_beg)-i_beg);
@@ -536,7 +536,7 @@ SSess::SSess( const string &iurl, const string &isender, const string &iuser, ve
     url(iurl), sender(isender), user(iuser), content(icontent), vars(ivars)
 {
     //> URL parameters parse
-    unsigned prmSep = iurl.find("?");
+    size_t prmSep = iurl.find("?");
     if(prmSep != string::npos)
     {
 	url = iurl.substr(0,prmSep);
@@ -564,7 +564,7 @@ SSess::SSess( const string &iurl, const string &isender, const string &iuser, ve
         }
     if(boundary.empty()) return;
 
-    for(unsigned pos = 0, spos = 0, i_bnd = 0; true; )
+    for(size_t pos = 0, spos = 0, i_bnd = 0; true; )
     {
         pos = content.find(boundary,pos);
         if(pos == string::npos || content.compare(pos+boundary.size(),2,c_end) == 0) break;
